requirements.txt
streamlit
polars
pycrypto